Keyword_comparative/requirements.txt
fastapi
uvicorn
sqlalchemy
psycopg2-binary
python-dotenv
pymysql
cryptography
pydantic-settings
pandas

